requirements.txt
AwesomeTkinter==2021.11.8
blinker==1.9.0
click==8.2.0
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.4.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
tzdata==2025.1
Werkzeug==3.1.3
